Staging (#13)
* Patch bugs

* Added locale for password validation and make monitored patient renders widget real time when new data is added

* Latest commit

* changes print to debugPrint

* Added new entry point for calculating mews

* Added dismiss for export page

* Modified alarm service for firebase cloud messaging

* Added citation

* Added citation

* Config alarm service

* Change package dependency for timezone

* Fix android build

* Upgraded gradle build for android

* Fix android build issue

* Added permission request for android

* Fixed permission request for android

* Added feature for notifying user even app is terminated

* Fix noti for ios, remove alarm when taking out patient of monitored state

* Bumped version

* Make Table calender works for report widget

* Add build runner to build apk script

* Bumped version for test ci

* test ci

* Edited Fastfiles

* added build runner in yml

* test ci 3

* test ci 4

* Test ci 5

* Test ci 6

* Test ci 7

* Test ci 8

* Added parallel pipeline

* Test ci

* Test ci

* Test ci

* Test ci

* Test ci

* Test ci

* Test ci

* Fix ios end

* Added production lane

* Fix match git URL for CI

* Add app identifier to Matchfile

+++ lib/services/logout_service.dart
@@ -14,9 +14,12 @@ class LogoutService {
   Future<Map<int, String>> logout() async {
     final secureStorage = SecureStorage();
     secureStorage.delete(key: 'session_cookie');
-    secureStorage.delete(key: 'nurseId');
-    secureStorage.delete(key: 'password');
-    secureStorage.delete(key: 'rememberMe');
+
+    if (await secureStorage.read(key: 'rememberMe') != 'true') {
+      secureStorage.delete(key: 'nurseId');
+      secureStorage.delete(key: 'password');
+    }
+
     await AlarmService().stopAllAlarms();
 
     await navigator.pushAndRemoveUntil(MaterialPageRoute(builder: (context) => const LoginPage()), (Route<dynamic> route) => false);
@@ -30,6 +33,7 @@ class LogoutService {
 
     try {
       final response = await http.get(url, headers: {"Content-Type": "application/json", "Authorization": "Bearer $idToken"});
+      secureStorage.delete(key: 'idToken');
 
       return {response.statusCode: response.body};
     } catch (e) {
